Fix scrolling problems on markdown transforms.
iOS7+ UITextViews have the following problem when `setText:` occurs:

1. `setText:` is called invalidating the text container.
2. CoreAnimation performs a rendering pass, triggering `layoutSubviews`.
3. `layoutSubviews` (or something before it) discards all the tile layers or the text container.  It then performs an invalid height calculation, returning something around 200, possibly the height of the scroll view.
4. `layoutSubviews` sets content size to the invalid value causing either the text view to scroll if scrolled too far or randomly triggers incorrect placement of the selection view.

The miscalculation is avoided by forcing the text view to render after performing `setText:`.  This can be done by rendering the text view's layer to a new context.  Calling `setNeedsLayout` appears to also be needed to trigger automatic scrolling to the selection if it is now partially offscreen.